+++ src/ui/ASelect/ASelect.js
@@ -489,7 +489,6 @@ export default {
 
     const {
       ariaLabelledby,
-      attributesDisabled,
       containerId,
       idForList,
       tabindex,
@@ -606,7 +605,6 @@ export default {
     return {
       ariaDescribedbyLocal,
       ariaLabelledby,
-      attributesDisabled,
       attributesToExcludeFromRender,
       buttonRef,
       clearModel,
@@ -730,7 +728,6 @@ export default {
               onKeydown: this.handleKeydown,
               onFocus: this.onFocus,
               onBlur: this.onBlur,
-              ...this.attributesDisabled,
             }, [
               this.$slots.fixedPlaceholder ?
                 this.$slots.fixedPlaceholder({